Various dashboard fixes after big b2g commit

+++ bin/cleanup-phone.py
@@ -13,8 +13,13 @@ def main(args=sys.argv[1:]):
     devicePrefs = eideticker.getDevicePrefs(options)
     device = eideticker.getDevice(**devicePrefs)
 
+    if device.type == 'b2g':
+        print "No cleanup presently required on b2g"
+        return
+
     for product in default_products:
-        device.killProcess(product['appname'])
+        if product['platform'] == device.type:
+            device.killProcess(product['appname'])
 
     # clean up any test stuff (profiles, etc.)
     device.removeDir(device.getDeviceRoot())
